requirements.txt
et-xmlfile==1.0.1
numpy==1.19.5
openpyxl==3.0.7
pandas==1.1.5
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
